Final build system for Android
