+++ Applications/Android/CameraModule/build.gradle
@@ -23,6 +23,11 @@ android {
                           "-DANDROID_CPP_FEATURES=rtti exceptions"
             }
         }
+
+        ndk {
+            //abiFilters 'x86', 'x86_64', 'armeabi', 'armeabi-v7a', 'arm64-v8a'
+            abiFilters 'armeabi-v7a'
+        }
     }
 
     compileOptions {
