--- Applications/Android/CameraModule/src/main/cpp/native-lib.cpp
@@ -1,12 +1,0 @@
-#include <jni.h>
-#include <android/log.h>
-
-#include <opencv2/imgproc.hpp>
-
-extern "C"
-{
-int testFunc()
-{
-    return 1;
-}
-}
