--- Applications/Android/Native/.gitignore
@@ -1,2 +1,0 @@
-libs/
-obj/
