--- Applications/Android/Native/Application.mk
@@ -1,4 +1,0 @@
-APP_STL := c++_static
-APP_CPPFLAGS := -std=c++11 -frtti -fexceptions
-APP_ABI := armeabi-v7a
-APP_PLATFORM := android-23
