--- Applications/Android/Native/_rebuild.bat
@@ -1,20 +1,0 @@
-rmdir /S /Q obj
-rmdir /S /Q libs
-
-if not exist jni mkdir jni
-
-del /F /Q jni\Dependencies.mk
-copy Dependencies.mk jni
-
-del /F /Q jni\Android.mk
-copy Android.mk jni
-
-del /F /Q jni\Application.mk
-copy Application.mk jni
-
-cd jni
-call %ANDROID_NDK%\ndk-build.cmd
-
-xcopy ..\libs\armeabi-v7a\libface_native.so ..\..\CameraModule\src\main\jniLibs\armeabi-v7a\ /D /Y
-
-pause
